feat(users/sterni/nix/buildC): pure nix build system for C
This is somewhat limited and needs to be fleshed out in the future.
Building shared objects and C++ could be supported in the future. Static
linking needs to be tested (I hope pkgsStatic.llvmPackages.stdenv just
works™).

The basic idea is based on buildLisp: We take a bunch of files and mash
them together. Changes in headers causes everything to be rebuild (which
is probably the only sane way to handle header changes except for
manually specifying dependencies which is annoying). Objects are built
individually, so rebuilding big projects should be cheap if headers are
not changed (to be tested).

Future work is mainly figuring nice support for testing (probably via
drvSeqL) and smoothing out the rough edges. Also figuring out when / if
this is ready to be put into //nix/buildC.

Change-Id: Ic24efe6b05eb761e19e4434f689c73ea2d075605