feat(nix/buildC): pure nix build system for C
This is somewhat limited and needs to be fleshed out in the future.

The basic idea is based on buildLisp: We take a bunch of files and mash
them together. Changes in headers causes everything to be rebuild (which
is probably the only sane way to handle header changes except for
manually specifying dependencies which is annoying). Objects are built
individually, so rebuilding big projects should be cheap if headers are
not changed (to be tested).

We support building shared and static objects, however static linking
requires more work: We need to a) propagate dependencies, probably by
recoursing into a tree and b) actually link binaries statically which is
hindered by -static -lc not working with default stdenvs.

Testing (probably via drvSeqL) will also have to be added.

Change-Id: Ic24efe6b05eb761e19e4434f689c73ea2d075605