feat(nix/buildC): pure nix build system for C
buildC is a build system for C libraries and programs plus attached test
suites implemented in Nix. Its main properties are that it builds every
object as an individual derivation, allowing nix to cache build results
and detect the need for rebuilds. It also works with the plain C compiler,
bintool and linker binaries instead of using the wrapper scripts found
in nixpkgs and stdenv.

This is achieved by making relatively strong assumptions over the
directory layout of the used and built libraries and passing around
additional meta data. We also don't support a lot of existing concepts
or tooling for building C libraries and programs, for example there is
no such thing as a configure phase in buildC.

There is some preliminary support for linking against libraries not
built using buildC in the form of buildC.libraryFromDrv. This is good
enough to support musl and glibc and has also successfully used to link
against simple libraries like libschrift and “heavier” ones like
HarfBuzz. At some point it may be necessary to use a mechanism like
pkg-config or implement something similar ourselves (to avoid
import-from-derivation). Generating .pc files could also be worthwhile.

Another feature is relatively great flexiblity: buildC supports both the
GCC and clang compiler toolchains, both musl and glibc (and
theoretically any other libc) and static as well as dynamic linking.
To prevent an unnecessarily large build and runtime closure all build
settings are applied to the entire dependency tree by overriding the
buildConf of every derivation depended on with the configuration of the
derivation being built. The main need for this is of course static
linking: This way we make sure that every library is built as a static
archive without requiring us to always build both a static and dynamic
version of the library. Additionally by preventing different libcs /
compilers being used we save on closure size both at run and build time
although you may need to rebuild more often if you use different
configurations for different derivations.

Of course this overriding mechanism doesn't work for buildC.libraryFromDrv,
but these can specify a predicate which indicates if they are compatible
with the current buildConf. For example, using buildC.glibc for a static
build will hit this predicate and fail at evaluation time.

Further work:

* Implement generic overriding support similar to buildLisp's overrideLisp
  or overrideAttrs

* Cross compilation support. In theory this already should almost work,
  but must be made convenient and requires some plumbing in the test
  code and likely runExecline.

* Collect more experience and fix upcoming issues: buildC works quite
  well for a limited set of usecases I personally had. Likely I'll
  encounter things in the future that should be supported, but aren't.

Change-Id: Ic24efe6b05eb761e19e4434f689c73ea2d075605