Update patch set 1
Patch Set 1:

Note that the zone doesn't actually delegate to these nameservers yet. Records are just configured in the domains.google UI.

The zone is already set up in GCP, but due to DNSSEC it will take a few days before we can move over (waiting for DNSSEC to be disabled and flushed out of all caches).

Patch-set: 1