feat(ops/nixos): Import Sourcegraph configuration into depot
Imports the Sourcegraph host configuration into config files generated
from Nix (don't worry - it's just builtins.toJSON).

The configuration was previously done in the Sourcegraph Admin UI,
which is not a great idea considering how the rest of our setup works.

The config is mostly imported exactly as-is, with the exception that
I've disabled `Batch Changes` (formerly `Campaigns`) because they
don't work with Gerrit anyways.

Change-Id: I1796dae96975919c94ff5e30b138b4d02e04b245