feat(sterni/dot-time-man-pages): mdoc port of dotti.me
More or less direct port of https://dotti.me to mdoc(7) with the
following changes:

* Add a RFC3339 column to the EXAMPLES table. RFC3339 is a well
  specified subset of ISO8601 whose specification is also more
  accessible so this could help someone out.

* Add a SEE ALSO section linking to the web site

* Add an AUTHORS section

The derivation is built using the custom, but fairly trivial
//nix/buildManPages builder which builds each man page, lints it using
mandoc (but only fails if requested to do so because (especially stable)
mandoc is pretty strict), gzips it and installs it to $out in the
correct directory structure.

Additionally nix.buildManPages.single is provided to build just one man
page where $out is the resulting man page.

Change-Id: I8db00bd402697aa52f6f651f28692617b487f832