fix(3p/lisp/puri): Use fetchgit for source, update version
Something changed in the upstream we fetch this source from that's
causing the fetch to fail - I can only assume it's a yanked rev, but
I'm not really sure. fetchgit from nixpkgs appears to be a little bit
more robust than builtins.fetchGit, so let's switch to that, and also
upgrade to a rev that we know is present.

Fixes: b/96
Change-Id: I8983c2df11ab4fa20f60915f950c6a7378efd2fd