feat(3p/cgit): limit width of div#summary, i. e. /about/**
Decreased text width for the /about pages should increase readability
considerably as jumping back to the beginning is hard for longer
lines. The result is still not perfect as the font size for the /about
pages is rather small and many lines thus get broken somewhat awkwardly.

We could probably migitate that using a larger font size.

The implementation choice of adding a tvl-extra.css which we inject into
cgit.css in preBuild is for simplicity: We don't need to worry about
routing an extra CSS file and loading it from the right location via
extra cgit head entries and serving it at the correct location using
either nginx or thttpd.

A drawback of this is however that iteration is slowed down by cgit's
compilation time.

Additionally, this should be the basis for implementing a bubblegum
themed cgit for Profpatsch.

Change-Id: I18060f735167acd623cef7a17c83408978461249