feat(3p/rust-crates): get dependencies from the depot fix point
I think it is good practice to always get dependencies from the depot
fix point if they are exposed. The reasoning for this is that if we
improve the support for overriding in depot, say by introducing a
depot.extend functions or even full blown overlay support, this will
already work as expected.

Change-Id: Ibb8dffcf32e8f46817a2db2da26139fabdce55bc