chore(3p/gerrit_plugins): attempt to stabilise hashes a bit
I'm dropping the leaveDotGit and deepClone bits; they were set like that
purely to try to make the build stamping work. In practice, not only
does the build stamping not work, but it also means we hit some
inconveniently-different hashes from time to time when gitiles does...
something??? on its backend.

I'm also putting some gcroots for these on whitby, which should also
help a bit, although it's a bit of a hack.

Change-Id: Ie6082248393e62795c18b1971fc2d16f4e8cc81d