chore(converse): Use naersk for building
This drops the old Carnix-generated file. Before moving back to
another builder I want to simplify the dependency tree of this.

Back when it was written, we had a stack that used Actix a lot - but
Actix has a ridiculous number of dependencies. We can trim this down a
bit.

Change-Id: I1dde9dc0325da7bdcb6608359fab33e27692dc1d