feat(3p/nixpkgs): apply haskell overlay only as a crossOverlay
crossOverlays are an undocumented feature of nixpkgs which allows us to
apply an overlay only to the “run packages” (that is pkgsHostTarget) and
not to buildPackages. This is useful for us since haskellPackages'
import from derivation tooling takes cabal2nix and the like from
buildPackages. Since we don't care about the haskell package set they
use as long as it isn't broken, it is enough to only apply our overlay
to the run haskell packages set. The benefit of this is that we should
be able to substitute everything we need at eval time from
cache.nixos.org.

Change-Id: I456a373f3d7116f10b7d9b4be2db3b41889fa4b7