feat(3p/nixpkgs): apply haskell overlay only as a crossOverlay
crossOverlays are an undocumented feature of nixpkgs which allows us to
apply an overlay only to the “run packages” (that is pkgsHostTarget) and
not to buildPackages. This is useful for us since haskellPackages'
import from derivation tooling takes cabal2nix and the like from
buildPackages. Since we don't care about the haskell package set they
use as long as it isn't broken, it is enough to only apply our overlay
to the run haskell packages set. The benefit of this is that we should
be able to substitute everything we need at eval time from
cache.nixos.org.

By setting crossOverlays to a non empty list however, we make nixpkgs
take the cross compilation code path for bootstrapping stdenv and the
package set. Most notably, this means for us that we loose
targetPackages.runtimeShell (it isn't available because we are cross
compiling in a sense). This breaks the evaluation of stuff like
pkgs.makeWrapper which we need to take from buildPackages.makeWrapper
instead.

stdenv.mkDerivation has some hacks to make it unnecessary to make this
work without explicitly taking packages from buildPackages, pkgs or
targetPackages. Depending on whether drv is in buildInputs or
nativeBuildInputs it chooses drv.nativeDrv or drv.crossDrv. However this
only works if drv has been passed to mkDerivation via pkgs.callPackage
which does some extra plumbing and sets these attributes. So
effectively, we can't use that behavior in depot and need to specify
buildPackages etc. manually where appropriate. This CL does a first step
towards this and cleans up some “wrong” imports where it would lead to
an evaluation error with crossOverlays != []. This does _not_ mean that
we could cross compile depot yet since there are probably additional
issues which would need to be resolved in a CL implementing cross
compilation.

This CL is presented as an alternative to cl/2921 to solve our issues
recompiling haskell dependencies when building the depot CI pipeline.

Change-Id: I456a373f3d7116f10b7d9b4be2db3b41889fa4b7