chore(3p/overlays): Remove LLVM11 bump from overlay
This was mostly for //third_party/nix and its dependencies which also
compiles with LLVM 7.1.0 it seems.

For //users/grfn/achilles we can manually select the newer LLVM version.

For the future we may want to have something similar again, but it may
not be necessary to invest too much time into it: nixpkgs is set to
upgrade their default llvmPackages to LLVM 11 as well at some point in
the near future.

Change-Id: Id83868dbc476a6c776b59518b856c933f30ea79d