feat(paroxysm): don't print error messages for question marks
If someone sends a message containing just question marks / spaces,
provided it has 2 question marks before it, paroxysm would attempt to
fetch an entry for the remainder of the question marks, and usually
fail. This fixes that oversight, and silences the "never heard of it"
message in such cases.

Joke entries created for such question mark cases will still fire,
though.

Change-Id: I44ef823a55c32869ab5f47ffc733ea566e23bb41
Reviewed-on: https://cl.tvl.fyi/c/depot/+/3161
Tested-by: BuildkiteCI
Reviewed-by: sterni <sternenseemann@systemli.org>
Reviewed-by: grfn <grfn@gws.fyi>
Reviewed-by: tazjin <mail@tazj.in>
Reviewed-by: cynthia <cynthia@tvl.fyi>