fix(tazjin/emacs): Temporary fixes to use unstable telega
The latest Emacs versions removed some (private) functions that telega
depends on, and this is fixed in HEAD of telega.el.

However, without these fixes, the unstable version of telega doesn't
build because the patch Nix tries to apply doesn't match the source
anymore.

The patch itself doesn't seem to do anything relevant for me.

Change-Id: Ib9a042c636cb438b2b15d231a07afd5c02be72ee